+++ ChangeLog.md
@@ -37,6 +37,7 @@ Full details of the variables available for each noted event, and VoiceAttack in
     * `Carrier jumped` revised to update `ContestingPowers` and `contestingpowers` and add `NearbyPowers`, `nearbypowers`, `powerAcquisitionProgress`, `powerControlProgress`, `powerReinforcementControlPoints`, and `powerUnderminingControlPoints` properties. 
     * `Carrier location` added, triggered at startup and shortly before your fleet carrier arrives at a new destination.
     * `Colonisation claim processed` added, triggered when staking or releasing a claim to colonise a star system.
+    * `Colonisation construction depot` added, triggered when progress is updated at the colonisation construction depot where you are docked.
     * `Colonisation contribution` added, triggered when contributing commodities towards construction in a colonised star system.
     * `Jumped` revised to update `ContestingPowers` and `contestingpowers` and add `NearbyPowers`, `nearbypowers`, `powerAcquisitionProgress`, `powerControlProgress`, `powerReinforcementControlPoints`, and `powerUnderminingControlPoints` properties. 
     * `Location` revised to update `ContestingPowers` and `contestingpowers` and add `NearbyPowers`, `nearbypowers`, `powerAcquisitionProgress`, `powerControlProgress`, `powerReinforcementControlPoints`, and `powerUnderminingControlPoints` properties.